requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
protobuf==5.26.1
requests==2.31.0
Werkzeug==3.0.2
gunicorn==22.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytest==8.3.3
pytest-flask==1.3.0